pyproject.toml
[tool.poetry]
name = "drf-api"
version = "0.1.0"
description = ""
authors = ["Batool Malkawi <batoolmalkawii@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.5"
djangorestframework = "^3.12.2"
pip = "^20.3.3"
install = "^1.3.4"
djangorestframework-simplejwt = "^4.6.0"
djangorestframework-jwt = "^1.11.0"
gunicorn = "^20.0.4"
whitenoise = {extras = ["brotli"], version = "^5.2.0"}

[tool.poetry.dev-dependencies]
black = "^20.8b1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
